+++ tests/testthat/test-shiny_general.R
@@ -91,7 +91,11 @@ local({
       dv.manager::mm_dispatch("unfiltered_dataset", "mpg"),
       "mod4"
     ),
-    "Name And Dataset" = dv.manager:::mod_dataset_name_date("mod_dataset_name_date")
+    "Name And Dataset" = dv.manager:::mod_dataset_name_date("mod_dataset_name_date"),
+    "Filtered Carb" = dv.manager:::mod_simple(
+      dv.manager::mm_dispatch("filtered_dataset", "carb"),
+      "mod5"
+    )
   )
 
   datasets <- list(
@@ -106,7 +110,6 @@ local({
     Sys.setenv("LC_TIME" = "en_US.UTF-8")
     do.call(dv.manager::run_app, !!args)
   })
-
   root_app <- start_app_driver(app_expr)
 
   # We store these values for using while testing
@@ -221,14 +224,13 @@ local({
   )
 
   test_that(
-    "active dataset can be filtered using the filtered menu" |>
-      vdoc[["add_spec"]](c(specs$active_dataset_filtering, specs$filter_key)),
+    "active dataset can be filtered using the global filter" |>
+      vdoc[["add_spec"]](c(specs$active_dataset_filtering, specs$filter_key, specs$global_filtering)),
     {
       skip_if_not_running_shiny_tests()
       skip_if_suspect_check()
 
       app <- shinytest2::AppDriver$new(root_app$get_url())
-
       # TODO: brittle too coupled with dv.filters
 
       app$set_inputs("global_filter-vars" = "car")
@@ -237,6 +239,28 @@ local({
       app$wait_for_idle()
       val <- app$wait_for_value(output = "mod1-text", ignore = list("4"), timeout = 10000)
       expect_identical(val, "1")
+    }
+  )
+
+  test_that(
+    "single data table from active dataset can be filtered using the single data table filter menu" |>
+      vdoc[["add_spec"]](c(specs$active_dataset_filtering, specs$single_filtering)),
+    {
+      skip_if_not_running_shiny_tests()
+      skip_if_suspect_check()
+
+      app <- shinytest2::AppDriver$new(root_app$get_url())
+
+      # TODO: brittle too coupled with dv.filters
+
+      app$set_inputs("dataset_filter_46ab8635-vars" = "carb")
+      app$wait_for_idle()
+      app$set_inputs("dataset_filter_46ab8635-carb" = c(1, 1))
+      app$wait_for_idle()
+      app$set_inputs("main_tab_panel" = "Filtered Carb")
+      app$wait_for_idle()
+      val <- app$wait_for_value(output = "mod5-text", ignore = list("4"), timeout = 10000)
+      expect_identical(val, "2")
     }
   )
 
@@ -383,7 +407,7 @@ local({
         # TODO: Split test
         expect_identical(
           val <- app$get_values(output = "mod_dataset_name_date-text")[["output"]][["mod_dataset_name_date-text"]],
-          "dataset_name: mpg_one_date ; dataset_date_range: 2021-01-13 2021-01-13 ; module_name: Filtered Tab,Returned Filtered,Read Output,Unfiltered Tab,Name And Dataset"
+          "dataset_name: mpg_one_date ; dataset_date_range: 2021-01-13 2021-01-13 ; module_name: Filtered Tab,Returned Filtered,Read Output,Unfiltered Tab,Name And Dataset,Filtered Carb"
         )
       }
     )
@@ -442,7 +466,7 @@ local({
     expect_equal(mpg_no_date[["current"]], mpg_no_date[["expected"]])
   })
 
-  test_that(vdoc[["add_spec"]]("filtering and dataset switching", c(specs$filtering_menu, specs$dataset_selector)), {
+  test_that(vdoc[["add_spec"]]("filtering and dataset switching", c(specs$filtering_menus, specs$dataset_selector)), {
     skip_if_not_running_shiny_tests()
     skip_if_suspect_check()
 
@@ -507,7 +531,17 @@ local({
 
     # Poor mans wait for accepted value
     app$set_inputs(main_tab_panel = "Filtered Tab")
-    app$wait_for_value(output = "mod1-text", ignore = as.character(setdiff(1:20, 6)), timeout = 10000)
+    app$wait_for_idle()
+    value <- local({
+      value <- app$get_values(output = "mod1-text")[["output"]][["mod1-text"]]
+      tries <- 10
+      while (!identical(value, "6") && tries > 0) {
+        tries <- tries - 1
+        Sys.sleep(1)
+      }
+    })
+
+    # app$wait_for_value(output = "mod1-text", ignore = as.character(setdiff(1:20, 6)), timeout = 10000)
     mpg_one_date_no_filter <- list(
       current = get_all_(),
       expected = list(
@@ -527,7 +561,15 @@ local({
 
     # Poor mans wait for accepted value
     app$set_inputs(main_tab_panel = "Filtered Tab")
-    app$wait_for_value(output = "mod1-text", ignore = as.character(setdiff(1:20, 4)), timeout = 10000)
+    app$wait_for_idle()
+    value <- local({
+      value <- app$get_values(output = "mod1-text")[["output"]][["mod1-text"]]
+      tries <- 10
+      while (!identical(value, "4") && tries > 0) {
+        tries <- tries - 1
+        Sys.sleep(1)
+      }
+    })
     mpg_one_date_filter <- list(
       current = get_all_(),
       expected = list(
@@ -543,7 +585,15 @@ local({
 
     # Poor mans wait for accepted value
     app$set_inputs(main_tab_panel = "Filtered Tab")
-    app$wait_for_value(output = "mod1-text", ignore = as.character(setdiff(1:20, 6)), timeout = 10000)
+    app$wait_for_idle()
+    value <- local({
+      value <- app$get_values(output = "mod1-text")[["output"]][["mod1-text"]]
+      tries <- 10
+      while (!identical(value, "6") && tries > 0) {
+        tries <- tries - 1
+        Sys.sleep(1)
+      }
+    })
     mpg_no_date_no_filter <- list(
       current = get_all_(),
       expected = list(
@@ -560,7 +610,16 @@ local({
 
     # Poor mans wait for accepted value
     app$set_inputs(main_tab_panel = "Filtered Tab")
-    app$wait_for_value(output = "mod1-text", ignore = as.character(setdiff(1:20, 4)), timeout = 10000)
+    app$wait_for_idle()
+    value <- local({
+      value <- app$get_values(output = "mod1-text")[["output"]][["mod1-text"]]
+      tries <- 10
+      while (!identical(value, "4") && tries > 0) {
+        tries <- tries - 1
+        Sys.sleep(1)
+      }
+    })
+
     mpg_no_date_filter <- list(
       current = get_all_(),
       expected = list(